requirements.txt
pyperclip
lru-dict